Use error estimates to map; add bootstrap mode
This is a possible scheme for using our dynamic error estimates to map.
Start by bootstrapping error data from the first N reads, optimized to
return from search as soon as a read is discerned not to be a unique
mapper, then map with the bootstrapped error estimates, updating the
estimates every N reads. While all tests are passing, I am not sure if
the actual error estimates are correct - need to check.