Unified statmap shared library
makefile builds statmap.so, a shared library of all the code in
statmap's source. Added getters and setters to the global variables in
statmap.c so they can be manipulated easily from ctypes. The new
statmap.so is a drop-in replacement for libtrace.so, and the other
indpendent shared libraries I was originally developing.