Fix bugs related to new SAM code
In src/mapped_read.c, there was a bug in the get next read code when the
rdb was mmapped - the mapped read wasn't properly freed and set to NULL
if we had reached the end of the database. This lead to an assertion
after the last mapped read was read.

Fixed a typo in a format string in sam.c progress logging.

Added an rm to the utilities makefile clean rule. Otherwise, the
dependencies list for the utility doesn't include the files from src
that it includes; when these are updated, it does not trigger a rebuild
of the corresponding utility.