Fix Ctypes alignment bug in mapped_read_db type
Using a pthreads type directly in the struct makes it impossible for us
to dereference it in Ctypes (although we can still pass around the
pointer wihtout problems). This was the cleanest fix - use a pointer to
the pthreads type, and dynamically allocate (same strategy already used
in rawread.c).

No need for unnecessary wrappers now.