Error correction based solely on observed error data
This version of Statmap performs a mapping bootstrap based on
mismatches, then uses that and subsequent observed error data to map. It
does not use the estimates from the sequencer at all.

Some tests are not passing:

1. repeat testing isn't working because the test does not produce any
unique mappers, so we can't bootstrap error data.
2. dirty read testing produces a mix of unmappable and nonmapping reads,
wheras previously it only produced unmappables. This may not be a
problem.