Use an explicit flag for ignoring results
We use this flag in the error data bootstrap to mark a set of
mapped_locations as invalid (because it is not a unique mapper).

This is a cleaner solution that simply setting the number of results to
zero, which could potentially cause memory leaks if the mapped_location
struct were changed to have any dynamically allocated members (wouldn't
be able to loop over the members and free them individually if we needed
to).