Implemeneted fwd and rev penalty array strategy
To deal with searching/rechecking normal and rev comp reads, creates a
reversed copy of the initially generated fwd penalty array and passes
both penalty arrays to the functions that use them. These functions can
then determine which array to use depending on the read they are
processing.

This strategy uses more memory but creates a nicer abstraction for the
penalty functions - they just take a penalty array, and don't need to
behave differently based on the strand.