Draft of utility to convert paired end read formats
Converts from "old format" (reads have /1 and /2 suffixes) to "new
format" (splits these into separate files, one for each end, and the
reads correspond in their positions in the files).

This version is naive and assumes that the reads are input in the proper
order - i.e., for each /1 read we see, the next /2 read is assumed to be
it's pair. This works in some common situations but could fail depending
on the input. A proper solution would be to build a dictionary keyed by
the read names (before the suffix).