Don't determine read end from read name
We require that the read end be known when populate_rawread_from_fastq
is called, since we cannot rely on paired end reads having the
appropriate suffix.

Modifies input file type hueristic code so it passes NORMAL as UNKNOWN
as the read end. UNKNOWN is no longer allowed, but the read end is
irrelevant to this code so NORMAL works as well.