Split rawreads into rawreads and reads (with subtemplates)
rawreads are now just a container for the data from input describing
reads (just fastq at the moment). They are used internally to create an
abstraction layer between reads and the FASTQ file format.

reads have a unique name and a list of subtemplates - single end reads
are initialized with a single subtemplate, paired end reads are
initialized with two (one for each end).