Binary search for matching mapped locations
This is for an optimization in searching for mapped locations to use to build
candidate mappings.

We want to compare the true start of mapped locations, or

    mapped_loc->location.loc - mapped_locs->subseq_offset

Since this information is stored in different structs, we cannot use bsearch
directly on these structures. Instaed, we build a temporary base mapped
location and a temporary set of current mapped locations to do the search.