Use stdlib bsearch for matching mapped locations
Add the mapped_location's strand to the compare function for mapped locations.
It was already using the chr index, which is why this was working before.

Construct a key mapped_location using the offset information so we can search
directly in the set of potential_matches, returning a pointer to the matching
mapped_location.