Standardize all cmp_ functions
Comparison functions (those that can be passed to qsort, bsearch, etc.) now
have a standard style across Statmap. The new style is to declare the cmp_
function with the specific types it expects to be passed, and then casts all
instances where it's used as a parameter to qsort/bsearch. This makes the
comparison function definitions much cleaner and easier to read/write/edit.