Add cmd line param for sequencer source of reads
If -i is not specified, Statmap will use the original guess_input_file_type
hueristic to guess the source of the reads. The -i parameter can be used to
manually specify the source of reads.

At the moment it is a rather inelegant mapping directly to the enum values in
input_file_type_t. Once we've stabilized everything, should come back and make
this interface nicer.