First step to optimize ps_loc handling
New strategy for handling pseudo locations when finding matches for mapped
locations is to
 1) expand the pseudo locations for any base location into the matching subset
    of mapped_locations for the base location.
 2) for potential matches, save memory by searching the pseudo locations
    without expanding them.

Add a unit test to check if pseudo locations are being properly handled when
there are multiple indexable subtemplates. This is necessary because expanding
the base locations is all that is required in the case where there is only one
indexable subtemplate.