Optimize psloc handling in matching mapped_locations
Handle ps_locs and normal locations seperately.

For ps_locs, find the contiguous block of pseudo locations in the matching
subset, and then binary search each one for matches.

For the normal locations, just do a binary search over the potential matches.
Since base_locs are expanded and a base_loc is never a pseudo location, this
will simply skip over all of the pseudo locations.