Write expanded diploid pseudo locations to disk
If a pseudo location is a shared diploid location, this adds code to the index
building that expands it into the two paternal and maternal diploid locations
before writing it to disk.

This works with the previous machinery for expanding pseudo locations and
diploid locations, since the unfinished part was handling diploid pseudo
locations. All tests pass.