Comment out iterative mapping code
I'm commenting out the iterative mapping code for now to more freely work on
the new mapped reads and gapped reads mapping.

I used preprocesser blocks to remove the calls to the iterative mapping code
and temporarily renamed the iterative_mapping source file so the compiler rules
will ignore it.