Some tests are passing with new mapped_read_t
New mapped reads code compiles without warnings, is clean in Valgrind, and
all of the tests up to the test_repeat_sequence_finding test are passing.

The failing test is interesting - it reports (500/100) reads mapped, indicating
that the marginal mapping code is working correctly, but then it only prints
100 entries in the SAM file. This is probably a bug in the SAM code, perhaps in
the new mapped_read_index interface.