Remove flag and revise building mapped reads
Mapped reads pseudo structure and interface for building mapped reads has been
revised to conform to the design from yesterday afternoon.

This code compiles (with warnings from a stub function) and runs without
errors. It was designed to only pass the single end tests because the
paired end tests will require additional work to replace the mapped_read_location
flag. The tests are not passing, however, because of removing the filtering code
that was previously part of building mapped reads. However, it appears that
this filtering code may have been masking previous errors, so I am committing
this before investigating further.