Generalize building cm's from match
Instead of special cases for building candidate mappings depending on the assay
type, we use an algorithm that checks the distance (normalized with the offset)
of mapped locations to determine if a candidate mapping should be added. If
the distance is non-zero, this implies there was a gap in the reference genome
between the mapped locations, so we add another candidate mapping to represent
this.

Note: we will need to revisit the rd_len and possibly the subseq_offset fields
on candidate_mapping once the rest of the gapped alignment code is written.
These values aren't being used right now, but they will need to be set
correctly to properly interpret the candidate mappings.

All perfect single end tests pass.