Generic (mostly) SAM writing
This new code entirely replaces the special-cased SAM writing code for
paired and unpaired reads. There are some locations where the code still
relies on a decision between whether the reads are single end or paired end
(in the case of determining the mate pair, I'm not sure if it's possible to
get around this).

This is not finished - the mate pair field generation needs to be finished.
Also need to check that some metadata on candidate_mapping corresponds to what
the new code expects.

Code compiles with only "unused" warnings. All tests pass up to
test_highly_repeated_paired_end_diploid, which is nearly all of them.