Fix diploid paired end bug
I forgot to include the chromosome check in the new paired end candidate
mapping joining code. This caused it to incorrectly pair some reads, resulting
in more than the number of expected mappings and causing the test to fail.

All test pass (except for a longstanding bug in the aggregate building in
simulate_chipseq.py).