Specialized binary search algorithm for matching mapped locations
Search implementation that returns either the first matching mapped location
in a set of matching mapped locations, or the first mapped_location with a
position greater than that of the key location, which could match given the
maximum reference insert size.

This implementation is not currently being used - before it can be used, need
to update the data structure that stores matches so it can properly store
multiple matches to a single base location.