Store index search results in a pointer array
Previously this used mapped_locations_container, an entirely unnecessary
structure in this case because the size of the container is static (equal to
the number of indexable subtemplates).

This is in preparation for removing mapped_locations_container entirely. It
was doing double duty as both the container of index search results and
mapped locations matching, but is unnecssary for the former and inadequate for
the latter.