Always use first indexable subtemplate's results as matching basis
The new multiple matching code (binary search, data structure, etc.) makes the
assumption that the base mapped location is always the 5'-most location in any
set of matched mapped locations. This is easy to ensure as long as two
conditions are met:

1) the 1st indexable subtemplate in the set of indexable subtemplates is the
   5'-most indexable subtemplate
2) we always use the search results from the 1st indexable subtemplate as the
   basis for matching.

This commit changes the function that chooses a base mapped_locations to
always use the first mapped_locations, satisfying the second condition. The
first condition is already met in build_indexable_subtemplates_from_read_subtemplate.