Add cigar string to candidate mappings
Instead of using multiple candidate mappings to represent a gapped alignment
for a single read subtemplate, we return to the original assumption that each
candidate mapping corresponds to one read subtemplate (and not a contiguous
region in the genome).

At the moment, this is only happening on the candidate mappings. We still need
to add the cigar strings to the mapped reads, and update the mapped reads and
SAM code to remove the different kinds of sublocations.

All tests pass (because they do not test the cigar strings, yet).