Naive recheck on joined candidate mappings
Removes the recheck on individual candidate mappings (not actually correct in
the paired end case) and replaces it with a recheck on joined sets of candidate
mappings. The recheck is the same likelihood ratio test that was used before.

The code has been refactored to perform this check after candidate mappings are
built, and to only pass known good sets of joined candidate mappings to the
mapped reads building code.

All tests pass.