Output nonmapping reads to fastq
The old code to write nonmapping reads to fastq was broken after implementing
the new mapped_read_t, which requires that a mapped read have at least one
mapping. Now that we explicitly store lists of the read ids of unmappable and
nonmapping reads, this code is greatly simplified.

All tests pass.