Modified parameters to make pseudo locations tests pass
The multiple indexable subtemplates for repeat sequences tests were far too
slow using the old parameters for REFERENCE_INSERT_LEN_MAX in config.h and
n_dups in the tests. In order for the test to actually test pseudo locations
handling, it just has to have n_dups (the number of times each location is
duplicated) >= PSEUDO_LOC_MIN_SIZE. Therefore, these tests are large enough
to test all of the pseudo locations code, but small enough to still run
quickly.

This is not a slowdown compared to the old code - the pseudo locations code
with one indexable subtemplate still runs just as fast. The problem here is
one of combinatorial explosion - matching indexable subtemplates across
pseudo locations creates an enormous number of candidate mappings.

This problem may be mitigated partially when I add code to consider the penalty
during matching as well, although the penalty over the gaps won't be considered
until we filter joined candidate mappings.

This needs to be tested on real data to see how it performs, especially in
repeat-heavy situations.