Bugfix in calculating mapped read size
Forgot to skip initial null pointers in joined mapped locations array when
calculating mapped read size. This was an omission from implementing the
filtering of joined candidate mappings.

This explains the sporadic errors I was seeing in the mutated reads tests.
Noticed it due to some Valgrind errors in the non multiple indexable
subtemplates tests, which I was checking for good measure.

All tests pass.