Filter mapped locations by cumulative penalty
Since the match penalty is the product of the marginal probabilities of the
mapped locations in the match, we know that if this probability is ever less
than the minimum match penalty, it will never be able to be valid. This would
be caught and filtered by filter_joined_candidate_mappings, but we can optimize
by ignoring these potential matches in the first place.