Bugfix setting min_match_penalty
min_match_penalty is a log probability, so it is expected to be a negative
number. Therefore, using -1 as the "unset" value that we test to determine
whether to set it to the default value is incorrect, and makes it impossible
for the user to use -1 for the min_match_penalty.