Match mapped locations for reverse complemented reads
Performs the reference gap check correctly for forward and reverse complemented
reads. 3' multiple indexable subtemplates tests pass.

Adds code to building cigar strings for candidate mapping to account for the
sequence in the read subtemplate before the first indexable subtemplate and
after the last, which we assume is matched. These lengths are now correctly
added to the first and last match entires in the cigar string.