Bugfixes
Fix (0/100) bootstrapped message. This was using mapped_cnt as the number of
reads bootstrapped, but it was not being updated unless a mapped read was
successfully built, which of course never happens during bootstrapping. Now
mapped_cnt is updated everytime at least 1 candidate mapping is found, so it
should be correct for both mapping and bootstrapping.

Fix occasional dirty reads test failures. This was actually being caused by a
bug in the non nonmapping reads printing code - after checking to see if a file
was empty, forgot to rewind the file pointer, so no nonmapping reads were ever
printed to fastq.

All tests pass.