Allocate memory efficiently for joining cm's
The previous algorithm repeatedly realloc'ed memory for the joined mappings
pointer array and the penalty array. This is inefficient and unnecessary, as we
can calculate the number of joined candidate mappings for both the single and
paired end cases easily. This code calculates the number of joined mappings and
then makes a single call to malloc to allocate the memory for joining cm's.