Rebuild penalty array after filtering
Filtering sets of joined candidate mappings out of the joined_mappings array
is achieved by setting their pointers to NULL, but it is less desirable to
use such a flag for the penalty array of floats. Therefore, after filtering,
we collect the filtered penalties into a new array, free the old array, and
set the penalty array pointer to the new array of filtered penalties.

Note that code before this had a serious bug in reporting penalties.