Bugfix - handle failed gapped mappings
If a proposed gapped mapping fails to produce any gapped mappings with a
penalty above the threshold, we need to discard the original candidate
mapping. This code handles both ungapped and gapped assays in a unified
manner.

This fixes an occasional bug where no gapped candidate mapping could be found,
and the original temporary gapped mapping was passed on to the joining code,
raising an assertion.