Handle reads with probes overlapping the intron in RNA-Seq unit test
At the moment, Statmap mostly ignores reads with index probes that happen to
overlap the intron (mapping these is future work). However, it is possible that
if one of the index probes only slightly overlaps the intron, the penalty score
will be low enough that Statmap will consider it mapping.

When this happens, the cigar string check in the RNA-Seq test will fail, and
the start location check may also fail, depending on how the probe overlaps the
intron. There is nothing we can do about this (for now), so we when checks
fail, we allow them (printing notices) if the originating fragment is known to
have an index probe overlapping an intron.