Add rev comp reads to RNA-Seq tests
Generate reverse complemented reads to try to map to the gapped genome,
converts sequence to reverse complement in the sequence check. Minor refactor
in tests.py to make reverse complement a function, so it's easy to import
into other scripts.

Reverse complement reads appear to be mapping if they do not overlap the intron
but do not map properly (causing an error) if they do.