Fix bugs in config parsing
Fixes bug that allows the selection of file type to "fall through" and not be
caught with an error if the min and max scores are not achieved.

The hueristic failed on Illumina 1.8 reads that had a min error score of 33 and
a max of 67. A better heuristic may be needed, or we may want to encourage
setting this explicitly.

Fix bug in explicit file type configuration code (mismatch between switch
statement and usage message).