Global variable for assay type
Now that we need to do CAGE-specific stuff, the assay type will be needed in
numerous locations in find_candidate_mappings. Rather than pass an additional
parameter through many layers of function calls, I made it a global variable.
Furthermore, it satisfies our requirement for global variables in Statmap -
it is set once, at the invocation of the program, and never changed.

All tests pass.