Untemplated Gs progress (unfinished)
This code compiles and runs, but the iterative mapping test (the only one
that actually tests the new code, since it sets the CAGE assay type) segfaults.
This is almost certainly due to re-using the old mappings structure, which I
did to avoid a memory leak - when it is reallocated, the old references will
be lost.

This is tricky to code, and may require something ugly like declaring a new
set of candidate mappings before the call to make_assay_specific_corrections,
so it can be freed after it has been used.