Build set of candidate mappings for assay specific corrections
Simplest solution to the previous problem - declare a set of assay specific
mappings generated by the corrections code, and append pointers to it onto
the filtered set of joined_mappings. This needs to be declared at a level
higher than I would like, but we need to be able to free this *after* the
pointers to the candidate mappings have been used to build the mapped read.

All tests pass.