Cleanup=True to prevent false test failures (for now)
If Cleanup=False, then the original source files generated for each test are
not removed after the test is completed. This is a problem because
map_with_statmap uses all available *.fa files by default to build the genome.
This can sometimes lead to tests appearing to fail - when really, what is
happening is reads are mapping to sequence in genome files from previous tests.
Until the tests can be rewritten to avoid this, this is the best default
setting for the Cleanup variable.