Fix corner case in update error data from candidate mappings
This code was (*very* rarely) segfaulting because find_seq_ptr would return
NULL. This should never happen because of the checks we do on the read location
when building candidate mappings, but it was possible because the code was
not considering the truth length of the mapped read after trimming.

All tests pass. CAGE reads that were causing a segfault
(CAGE_AdMateM_4d_testes_rep1_535_cage_BS272.trimmed.fastq) have passed the
point in mapping where they were previously segfaulting.