Properly recheck reverse complemented gapped reads
The rewrite of recheck_candidate_mapping's internal recheck loop during work for
the CAGE release did not handle reverse complemented gapped reads properly.

Statmap's internal representation of a mapping (in a candidate_mapping) is the
same as the entry in a SAM file: the start_bp is the leftmost mapping position
(really the end of a reverse complemented read), and the cigar string is read
from left to right starting at start_bp. Since we now reverse complement the
entire genome fragment, we need to do this recheck from the read's true start,
so reverse complemented reads require an adjustment.

It may be possible to avoid this by reverse complementing the read instead.