Update the recheck in the RNA-Seq test for r.c. reads
We merged a fix to print the reverse complement of a sequence for any reads
that map to the reverse complement in SAM during work for the CAGE release.
This requires us to modify the recheck loop in the test (fortunately, it
actually allows us to simplify it significantly).

All tests pass. RNASeq test passes.