Lower bound min_match_penalty for CAGE
This guarantees that we can at least map perfect reads with up to the maximum
number of untemplated G's. This is a rough fix and the error model needs to be
reexamined to handle this properly.

Untemplated G's test is successfully bootstrapping and mapping all reads.

All other tests pass.