Fix single-threaded iterative mapping exiting prematurely
This bug was introduced in an earlier refactor of the update_trace code.
It tried to re-use the thread worker function for both the single threaded
and multi-threaded cases, but due to the call to pthread_exit in the worker
would exit prematurely and cleanly when doing iterative mapping with 1 thread.

Addded a thread-specific wrapper function around the worker that handles
calling pthread_exit.

All tests pass. Valgrind is clean.