Fix penalty computation when building gapped candidate mappings
The penalty computation was a holdover from the original algorithm, which
kept the original index probes and extended the exons from them. The new
algorithm recomputes the entire penalty, so leaving in code that adds
the candidate mapping's penalty was incorrectly making the penalties too low.

With this fix, we were able to map a 41bp subread from one of the RNA-Seq
reads. The mappings corresponded nicely to the top results from a BLAST
search for the same sequence.