Increase match stack's maximum length
Running Statmap on the real RNA-Seq data caused the match stack to overflow
its original limit of 500. So far it is mapping successfully with this new
limit.

A better solution would be to either empirically try to determine the upper
bound for the match stack, or use a dynamically allocated match stack.