Fix corner case bug in sequence match check
Fixes error when mapping coordinate overran the genome boundary due to soft
clip simulation. This was the cause of the occasional test failures in the
softclipped reads test.

Previously implemented this improved algorithm in simulate_rnaseq, but did not
update the similar function in tests.py. This is a cleaner way to compare a
mapped read's sequence back to the original genome sequence, considering CIGAR
ops M,N,S.