Convert penalty_t to a 4x1 matrix of floats
Since penalty arrays are built for specific read sequences, the observed base
pair at any position is fixed. Therefore, we only need to store the four
penalties of this base pair matching against any other base pair. This may
provide an optimization in the penalty array lookup in compute_penalty, which
is heavily used according to gprof.

This required a rewrite of the reverse penalty array code to build the reverse
array directly from the reverse sequence, and careful consideration of how
the reverse penalty array is handled in both the index search and the recheck.

In the index search, we needed to adjust the subseq offset for the reverse
penalty array since the subseq offset is given as the offset from the 5' start
of the read.

In the recheck, we actually reverse complement the mapped genome sequence given
by the cigar string to make the algorithm simpler and more efficient in the
gapped case. In this case, we only want to use the fwd penalty array, since we
are comparing the original read sequence against the genome sequence that it
mapped to.