Re-enable 3' test for multiple indexable subtemplates
This test was previously failing for odd-length reads that were reverse
complemented, because the second subtemplate's end rounded down would be 1
away from the true end of the read. This created an off-by-1 error in the
candidate mapping's start bp which would cause the read to be thrown out
in the recheck.

This solution works for now, but demonstrates that the current matching
code requires that the two subtemplates be at either end of the read. A
better solution would be to relax this requirement and allow them to be
anywhere in the read.