Copy a config.dat file to the merged output directory
We just need this so the post-processing scripts can figure out the original
assay type. In fact, most of the data (except for the genome_fname and the
assay_type, which we explicitly check as criteria for matching) in this file
will be erroneous. Many of these files have no correct setting in this
merged case.