Refactor rd_type field in candidate_mapping
This type was mostly leftover from an earlier time when candidate mappings
were supposed to represent contiguous portions of the genome (i.e., a gapped
read would be built from multiple candidate mappigns). All we need now is
the reference to the originating read subtemplate.

There may be further opportunity for refactoring the pos_in_tempalte field in
the read struct as well.