Bugfix for assay-specific corrected mappings close to contig boundary
When reads are soft-clipped for CAGE assay specific corrections, the
resulting set of corrected mappings could have invalid locations if the
original mapping was within MAX_NUM_UNTEMPLATED_GS of a contig boundary.
This is fixed by re-checking the validity of the locations of assay
specific corrected mappings.

This bug was discovered through an assertion raised in mapping
'CAGE_L1_284_cage_BS107.trimmed.fastq'. A previous commit partially
fixed this bug, but failed to handle the case where the mapping is
reverse complemented. This commit introduces a complete fix for all
possible cases.

All tests pass, as does mapping of the above fastq.