Fix bug caused by loading leftover genomes
Explicitly passes the genome_fnames generated for this test to
map_with_statmap to avoid the default behavior of treating any *.fa file
as a genome file. This was causing an occasional segfault in combination with
the new gap criteria because some mapped locations would appear to be
overlapping because they mapped to two different chrs that had the same name.