Build arbitrary # of indexable subtemplates
Simple strategy to build an arbitrary number of index probes from a read -
splits the read into equal sized partitions and uses the optimal subsequence
from within each partition as the probe. At the moment it uses at most 2
index probes, so it is equivalent to the previous hard-coded algorithm's output
(easily modified).

Fixes handling of softclipping - should be checked and adjusted for the
entire set of index probes, not each index probe. This worked for the CAGE
because there was only ever one index probe.