Fix sign of recheck_max_penalty_spread
Max penalty spread parameter should be positive - when we switched to taking the
log of the mapping metaparameter (which is 0 <= metaparameter <= 1) we got a
negative number. The result was the max_penalty_spread parameter was (silently)
ignored in both the index search and recheck.

This commit fixes the bug and adds asserts to check its value when it is set.