Log separately to logfile and stderr
Now logs to stderr if log level is above a certain level, and logs all
messages out to an external log file. Since we want to save the log file
in the Statmap output directory, we start by logging to a file in /tmp,
then copy what's been logged so far and re-open the log file in the output
directory once it's been created.