Probability of any mismatch is equally likely (for now)
The current model treats P(match) + P(mismatch) = 1. This means we can't sample
from the probability distribution over the penalties for each mutated base,
because they do not sum to 1. This is a temporary fix to make the probability
distribution correct until supported is added for different bp mutation rates
in the error model.