Fix for computing error_prb when 'N' in read sequence
Previously, ref would never == obs in error_prb_for_estimated_model,
so the resulting probability distribution would be incorrect. This
was raising assertions with the new min match penalty sampling code.

This is the identical stategy that we use in building the sequences
to search for in the index (replacing 'N' with 'A'). recheck_penalty
will actually consider the true penalty of an N in the read sequence