Scaling factor for index probe min_match_penalty
Use the ratio of the perfect match penalty for the index probe to the perfect
match penalty for the read to compute the index probe min match penalty from
the read (block) min match penalty.

This scaling factor should hopefully control for inhomogeneity in the
penalties; however, the RNASeq test is still failing.

Refactors the init_..._params functions to return a pointer instead of needing
a reference parameter, which was unnecessary.

Fixes bug in subseq_penalty in the penalty computation for loop's bounds. This
initially caused test_error_rate_estimation to fail, because the second index
probe's min match penalty was always 0, because the second probe's perfect
match penalty was always 0 (due to the bug).