Fix mutex design to give each segment a lock on the data
This design does not change any of the segment's metadata after it has been
created - just the values in the trace data. Therefore, other threads can
iterate over the array of segments to find the segment they need to update,
without bothering with checking the locks. It is only when a thread needs to
update a segment's data that it needs to check the lock. If, as we hope, each
read maps uniquely into a segment, then this should be an optimal concurrency
model.