Rewrite trace Python API to use new trace_t
Utilities that use the trace API, such as build_mle_trace.py, execute without
problems. Further testing is needed to make sure the output is as expected.

Also rewrites the Python trace_t type (to provide a more convenient way to
access trace data in Python scripts). This is currently unused, so the rewrite
is untested.