Add read fragment type to reads
When a read is loaded from the rawread db, determine its fragment type based on
its assay and whether it is single or paired end. This will later be used to
determine how to incorporate penalties from the fragment length distribution
into the rechecked penalty calculations.

Removes previous enum, is_full_fragment, that was intended to express this but
was not being used (and was not expressive enough).

Misc. cleanup