Join candidate mappings based on number of read subtemplates
Now we join based explicitly on whether the read is single or paired end.
The code was previously trying to determine whether a read was paired or not,
but this code was buggy. If we are processing paired reads, and a paired
candidate mapping cannot be built, the mapped read should not be built - we
should not instead build a single end mapped read.

simulate_chipseq now builds paired instead of single end reads (since we have
temporarily removed the single ended chipseq iterative mapping code). This
bug would occasionally come up when running simulate chipseq and confusingly
lead to a fatal error because it would try to iteratively map a single ended
chip read.