Fix assertion if number of reads is a multiple of READ_STAT_UPDATE_STEP_SIZE
This is caused by the slightly tricky problem of determining the EOF of the
rawread_db. rawread_db_is_empty uses feof, which only reports EOF if a
failed read has been made. If the number of reads is a multiple of
READ_STAT_UPDATE_STEP_SIZE, we will have processed all the candidate mappings
but not have any way to know it until we try to process another one. This
used to just fall through, but now is detected and the loop is explicitly
terminated using the return value if we are using the estimated model;
otherwise just falls through.