Basic trace segmentation function
Returns a list of trace segments that are at least MIN_TRACE_SEGMENT_SIZE bp's
long and are split whenever the underlying trace is.

Added code to build and print the segment list to sample_random_trace to check
that it's working. The actual segmented trace is not built/used. The original
trace (that is used to determine the trace segments)_is also not properly
initialized (I'm not sure what the right way is).