Avoid segfault if we try to log before logging has been initialized
Initializing logging should always be the first thing Statmap does, I realized
a segfault will occur (no log_fp or log_mutex initialized) if a logging call is
made before this happens while testing the new exit behavior for a fatal error.

Writing out to stderr is a fine failsafe, and ensures that the log message will
be communicated to the user.