Use R_tryEval to catch errors the occur in R
If an error (including an interrupt) occurs while R is executing from a call
to eval(), it tries to return to R's read-eval-print loop - not what we want
in an embedded setting. It creates really useless error messages, brings up
a pointless menu that doesn't do anything, and is a nightmare if something
goes wrong in an automated script (just spews the read-eval-print loop into
the log files).

This code allows us to detect if an error occurred while R was running. If
an error in R occurs, we simply report a fatal error (since we can't easily
inspect what's happening in R and try to recover). Also, this happens due to
keyboard interrupts 99% of the time anyway.