Avoid creating trace segments with length less than the min
New trace segmentation algorithm looks ahead to see if another trace can be
built after the current one; if there are not enough bp's, it extends the
current trace to the end. Also adds a special case in case the contig is
shorter than the minimum trace segment length - adds the segment, but warns
the user that it is doing so.