Revise trace segment search algorithm
This new algorithm is used both when updating trace functions and in building
the trace segment graph (it is repeated, unfortunately, due to the search
contexts being different).

This fixes occassional incorrect results from the graph building algorithm, and
occasional valgrind errors caused by the incorrect handling of a boundary
condition (when element sought is the last in the list).