Add check for mapped reads that overlap multiple trace segments
At this point, this is causing both the CAGE and Chipseq iterative mapping
tests to fail. Theoretically, we would only expect the CAGE test to fail
because when the initial trace is updated, it only updates the promoter
bp position and not the full fragment.

Adds todo item for bugs/issues with trace segmentation algorithm.