Clean up and have git ignore log file from build_index.py
Since build_index.py uses the Statmap shared library, and there is a call to
initialize logging in the constructor for StatmapOutput, a log file is created
in the working directory (tests). The log file code always appends to the log
file, so this file will continue to grow as `make check` is run repeatedly.

Adds rm to remove the old log file every time make check is run, and a rule to
git ignore the log file.